multi: Add more stringent checks to sched_group_set_rt_runtime
When setting the runtime of a whole cgroup, schedulability analyses are run to comfirm that tasks are not starved. If setting the runtime of a group to zero, and it has running tasks, of course tasks will starve. The original function shuts down each server one by one and then fails on shutting down the last server, this is wrong.