Update fair server initialization and bandwidth accounting
fair_dl_server_apply_params is used to initialize the fair_servers. Fair servers should
not be accounted in the bandwidth for real-time tasks, rather they are assigned to the
remainder of the bandwidth not assigned to real-time tasks. The fuction does not make
use of __dl_overflow, i.e. the bandwidth accounting check for real-time tasks.
Additionally, the fair server has assigned the correct remainder bandwidth instead of
the fixed 5%.