sched/rt: Disable RT_GROUP_SCHED
Disable the old RT_GROUP_SCHED scheduler. Note that this does not
completely remove all the RT_GROUP_SCHED functionality, just unhooks it
and removes most of the relevant functions. Some of the RT_GROUP_SCHED
functions are kept because they will be adapted for the HCBS scheduling.

Most notably:

- Disable the initialization of the rt_bandwidth for group scheduling.
- Unhook any functionality for RT_GROUP_SCHED in normal rt.c code, leaving
  only non-group functionality.
- Remove group related field initialization in init_rt_rq().
- Remove all the unhooked (and so unused) functions from RT_GROUP_SCHED.
- Remove all allocation/deallocation code for rt-groups, always returning
  failure on allocation.
- Update inc/dec_rt_tasks active tasks' counters, as rt scheduling
  entities now only represent a single task, and not a group of tasks
  anymore.

Signed-off-by: Yuri Andriaccio <yurand2000@gmail.com>