sched/rt: Remove support for cgroups-v1
Disable control files for cgroups-v1, and allow only cgroups-v2. This
should simplify maintaining the code, also because cgroups-v1 are
deprecated.

Set the default rt-cgroups runtime to zero, otherwise a cgroup-v1 kernel
will not be able to start SCHED_DEADLINE tasks. The bandwidth for
rt-cgroups must then be manually assigned after the kernel boots.

Signed-off-by: Yuri Andriaccio <yurand2000@gmail.com>