multi: Reintroduce __rt_schedulable_multi and tg_rt_schedulable_multi
The function is used once again, instead of __rt_schedulable which takes the average of runtimes.
A number of changes are done, since dl_se(s) of a task group cannot be uninitialized. The check when
changing the runtime of the root_task_group is now done under rcu_lock, like in __rt_schedulable.