sched/core: Execute enqueued balance callbacks when migrating task betweeen cgroups
Execute balancing callbacks when migrating task between cgroups, since
the HCBS scheduler, similarly to the previous patch, may request balancing
of throttled dl_servers to fully utilize the server's bandwidth.

Introduce the RELEASE_LOCK helper macro to explicitly unlock a guard-based
lock. The macro calls the destructor function of the lock and invalidates
it so that the lock is not unlocked twice when the lock variable goes out
of scope.

Signed-off-by: Yuri Andriaccio <yurand2000@gmail.com>