sched/rt: Allow zeroing the runtime of the root control group
This runtime only affects the available bandwidth of the rt-cgroup hierarchy but
not the SCHED_FIFO / SCHED_RR tasks on the global runqueue.

Disabling the root control group bandwidth should not cause any side effect, as
SCHED_FIFO / SCHED_RR tasks do not depend on it since the introduction of
fair_servers.

Signed-off-by: Yuri Andriaccio <yurand2000@gmail.com>