sched/rt: Allow zeroing the runtime of the root control group
In HCBS, the root control group follows the already existing rules for rt-task
scheduling. As such, it does not make use of the deadline servers to account for
runtime, or any other HCBS specific code and features.

While the runtime of SCHED_DEADLINE tasks depends on the global bandwidth
reserved for rt_tasks, the runtime of SCHED_FIFO/SCHED_RR tasks is limited by
the activation of fair-servers (as the RT_THROTTLING mechanism has been removed
in favour of them), thus their maximum bandwidth depends solely on the
fair-server settings (which are thightly related to the global bandwidth
reserved for rt-tasks) and the amount of SCHED_OTHER workload to run (recall
that if no SCHED_OTHER tasks are running, the FIFO/RR tasks may fully utilize
the CPU).

The values of runtime and period in the root cgroup's cpu controller do not
affect, by design of HCBS, the fair-server settings and similar (consequently
they do not affect the scheduling of FIFO/RR tasks in the root cgroup), but they
are just used to reserve a portion of the SCHED_DEADLINE bandwidth to the
scheduling of rt-cgroups. These values only affect child cgroups, their deadline
servers and their assigned FIFO/RR tasks.

Signed-off-by: Yuri Andriaccio <yurand2000@gmail.com>